Implements better loggin
- If applied, this commit will...
Instead of logging to console. Log to a file. When the log gets to be 1000
lines, it will get removed.

- Explain why this change is being made...
Noticed during debug mode, some programs are not happy with lots of lines being
dumped out. Happened while in spacemacs. Thus, logging to a file is better.
Reason for putting a limit on the file is because there are two sources being
run with shell and are both going to one log file. Letting the log file grow to
1000 lines was necessary to allow for both actions to be appended to one file.

- Provide links to any relevant tickets, articles, or other resources...