set the LIBSH_HOME if not set
This allows for setting the LIBSH_HOME if its not set. Needed when using libsh
from another script. Turns out that detecting the path from within a sh script
was not working. So I came up with the pattern of adding this to a script if you
want to use libsh inside a script, add the following:

```
LIBSH_HOME=path/to/libsh
source $LIBSH_HOME/libsh.sh
```