Add instructions for starting jupyterlabs with libsh
This is really early stage. The idea is that its customized from a common
starting point. My idea is that there should be a base script of packages
installed but there needs to be a balance. Too many packages makes the container
image too large.

By using a function it can be customized for the project your working on.