Adding aws v2 support
This change applies support for AWS v2. Turns out the new utility takes
a different method for installation and by default sends all output to
a pager for searching the output. v2aws relies on an output and having
to quit a pager every time seemed annoying. So I put in a option for
disabling the pager.